requirements.txt
kafka-python
pandas
transformers~=4.30.2
torch
xformers
confluent_kafka